+++ examples/filesystem-snapshots/CHANGELOG.md
@@ -1,5 +1,33 @@
 # sandbox-filesystem-snapshots
 
+## 0.0.15
+
+### Patch Changes
+
+- Updated dependencies [[`9dc0ac96d1d531b5a44786c92d6ddc2ce4890791`](https://github.com/vercel/sandbox/commit/9dc0ac96d1d531b5a44786c92d6ddc2ce4890791), [`494c2ddae5899b7f793cbd906a2a62260bd08885`](https://github.com/vercel/sandbox/commit/494c2ddae5899b7f793cbd906a2a62260bd08885), [`3fbabb95946771e41d8b17b7f8cb8d2972beba25`](https://github.com/vercel/sandbox/commit/3fbabb95946771e41d8b17b7f8cb8d2972beba25), [`42515e1ede4468fad204e5332ade0ee5d8dce3e5`](https://github.com/vercel/sandbox/commit/42515e1ede4468fad204e5332ade0ee5d8dce3e5)]:
+  - @vercel/sandbox@1.10.0
+
+## 0.0.14
+
+### Patch Changes
+
+- Updated dependencies [[`9555162f33690dfa18530aeca93af05188ebd2ed`](https://github.com/vercel/sandbox/commit/9555162f33690dfa18530aeca93af05188ebd2ed)]:
+  - @vercel/sandbox@1.9.3
+
+## 0.0.13
+
+### Patch Changes
+
+- Updated dependencies [[`b91b9e49fb7d2c5a4b601c07c125e6e5a2c43441`](https://github.com/vercel/sandbox/commit/b91b9e49fb7d2c5a4b601c07c125e6e5a2c43441)]:
+  - @vercel/sandbox@1.9.2
+
+## 0.0.12
+
+### Patch Changes
+
+- Updated dependencies [[`cf13a34221c2b83c25c73d94929d05e0a697aecf`](https://github.com/vercel/sandbox/commit/cf13a34221c2b83c25c73d94929d05e0a697aecf), [`772989c59a3c27efa98153cdc54b6e35c1c15eae`](https://github.com/vercel/sandbox/commit/772989c59a3c27efa98153cdc54b6e35c1c15eae), [`184cd42d8d3b1ea1df354529cb6ba103a33e18d3`](https://github.com/vercel/sandbox/commit/184cd42d8d3b1ea1df354529cb6ba103a33e18d3), [`451c42efb94ab9c9dc330b4742071ac01008044d`](https://github.com/vercel/sandbox/commit/451c42efb94ab9c9dc330b4742071ac01008044d)]:
+  - @vercel/sandbox@1.9.1
+
 ## 0.0.11
 
 ### Patch Changes